requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0

gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
reportlab==3.5.68
sqlparse==0.4.1
whitenoise==5.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-rest-auth==2.1.11

django-ckeditor==6.1.0
django-cors-headers==3.8.0


django-js-asset==1.2.2
django-rest-knox==4.1.0
django-telegrambot==1.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
drf-yasg==1.20.0
environs==9.3.3
et-xmlfile==1.1.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
Jinja2==3.0.1
jsonschema==2.6.0
ngrok==0.0.1
oauthlib==3.1.0
openpyxl==3.0.7
pathlib==1.0.1
Pillow==8.2.0
pyOpenSSL==20.0.1
PyRect==0.1.4
python-dotenv==0.17.1
python-jsonrpc-server==0.3.4
python3-openid==3.2.0
PyYAML==5.4.1
regex==2021.4.4
reportlab==3.5.68
requests-oauthlib==1.3.0
yarl==1.6.3